+++ demo/src/main/resources/templates/index.html
@@ -89,10 +89,10 @@
     <div class="container">
         <h1>🚀 Spring Boot 서버</h1>
         <p class="message" th:text="${message}">서버가 정상적으로 실행 중입니다!</p>
-        
+
         <div class="btn-group">
-            <a href="/login" class="btn btn-primary">로그인</a>
-            <a href="/signup" class="btn btn-secondary">회원가입</a>
+            <a href="login" class="btn btn-primary">로그인</a>
+            <a href="signup" class="btn btn-secondary">회원가입</a>
         </div>
 
         <div class="api-info">
